moved create_metricprofile_history function to history_helpers file
